Proof of second (obscure) conserved quantity
This quantity roughly equates to each pair of BA and AB gliders, as they alternate between being separate, or being merged as a cluster of As of Bs. This proof is very messy and might be hard to understand, but I can't find a more elegant way of formalizing what's in my head, so I went for an exhaustive method.

I realize that it will also be very hard to check whether this is accurate, but I'll try to simplify the proof at a later point.